geox/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arifos-geox"
version = "0.4.0"
description = "GEOX — Governed, agentic geological intelligence coprocessor for arifOS. DITEMPA BUKAN DIBERI."
readme = "README.md"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Muhammad Arif bin Fazil", email = "ariffazil@gmail.com"},
]
requires-python = ">=3.10"
keywords = [
    "geology",
    "agentic-ai",
    "mcp",
    "earth-model",
    "governed-ai",
    "arifos",
    "geoscience",
    "constitutional-ai",
    "hydrocarbon",
    "prospect-evaluation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "httpx>=0.24",
    "pyyaml>=6.0",
    "python-dateutil>=2.8",
    "numpy>=1.24",
    "pillow>=10.0",
    "scipy>=1.11",
    "scikit-image>=0.22",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
qdrant = [
    "qdrant-client>=1.7",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.7",
    "types-PyYAML>=6.0",
    "types-python-dateutil>=2.8",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.4",
    "mkdocstrings[python]>=0.24",
]

[project.urls]
Homepage = "https://github.com/ariffazil/GEOX"
Documentation = "https://github.com/ariffazil/GEOX#readme"
"Bug Tracker" = "https://github.com/ariffazil/GEOX/issues"

[project.scripts]
geox = "arifos.geox.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["arifos"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests that require external services",
    "slow: marks tests that run slow",
]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in default arguments
]

[tool.ruff.isort]
known-first-party = ["arifos"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["qdrant_client.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["arifos/geox"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 80
